Fixed spec for sort conditions.
`arachne.aristotle.query.compiler/op` expects queries to generally have at most three arguments. For
this to be true with `:top-n` and `:order` queries, the sort conditions need to be packed inside a
list.

With the old spec, the sort conditions were not nested, and thus `op` received the sort condition
keyword as `a2` and the direction keyword as `amore`, with the query appended.